Refactor app actions handler function into mixin for reuse
