+++ src/js/components/AppPageControlsComponent.jsx
@@ -2,20 +2,19 @@ var classNames = require("classnames");
 var OnClickOutsideMixin = require("react-onclickoutside");
 var React = require("react");
 
+var AppActionsHandlerMixin = require("../mixins/AppActionsHandlerMixin");
 var AppStatus = require("../constants/AppStatus");
-var AppsActions = require("../actions/AppsActions");
-var DialogActions = require("../actions/DialogActions");
-var DialogStore = require("../stores/DialogStore");
-var QueueActions = require("../actions/QueueActions");
 
 var AppPageControlsComponent = React.createClass({
 
   displayName: "AppPageControlsComponent",
 
-  mixins: [OnClickOutsideMixin],
+  mixins: [
+    AppActionsHandlerMixin,
+    OnClickOutsideMixin
+  ],
 
   propTypes: {
-    appId: React.PropTypes.string.isRequired,
     model: React.PropTypes.object.isRequired
   },
 
@@ -43,73 +42,6 @@ var AppPageControlsComponent = React.createClass({
   handleClickOutside: function () {
     this.setState({
       isDropdownActivated: false
-    });
-  },
-
-  handleResetDelay: function () {
-    QueueActions.resetDelay(this.props.appId);
-  },
-
-  handleScaleApp: function () {
-    var props = this.props;
-
-    const dialogId =
-      DialogActions.prompt("Scale to how many instances?",
-          props.model.instances.toString(), {
-            type: "number",
-            min: "0"
-          }
-      );
-
-    DialogStore.handleUserResponse(dialogId, instancesString => {
-      if (instancesString != null && instancesString !== "") {
-        let instances = parseInt(instancesString, 10);
-
-        AppsActions.scaleApp(props.appId, instances);
-      }
-    });
-  },
-
-  handleSuspendApp: function (event) {
-    event.preventDefault();
-
-    var props = this.props;
-
-    if (props.model.instances < 1) {
-      return;
-    }
-
-    const dialogId =
-      DialogActions.confirm("Suspend app by scaling to 0 instances?",
-        "Suspend");
-
-    DialogStore.handleUserResponse(dialogId, () => {
-      AppsActions.scaleApp(props.appId, 0);
-    });
-  },
-
-  handleRestartApp: function () {
-    var appId = this.props.appId;
-
-    const dialogId =
-      DialogActions.confirm(`Restart app '${appId}'?`, "Restart");
-
-    DialogStore.handleUserResponse(dialogId, () => {
-      AppsActions.restartApp(appId);
-    });
-  },
-
-  handleDestroyApp: function (event) {
-    event.preventDefault();
-
-    var appId = this.props.appId;
-
-    const dialogId =
-      DialogActions.confirm(`Destroy app '${appId}'? This is irreversible.`,
-        "Destroy");
-
-    DialogStore.handleUserResponse(dialogId, () => {
-      AppsActions.deleteApp(appId);
     });
   },
